+++ CHANGELOG.md
@@ -43,6 +43,8 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 - [#8](https://github.com/ARGOeu/argo-mon-status-api/pull/8) ARGO-5156 ARGO-MON-STATUS-API: Downgrade project postgres version to 11
 - [#25](https://github.com/ARGOeu/argo-mon-status-api/pull/25) ARGO-5190 Fix website in tenant to be empty #25
 - [#26](https://github.com/ARGOeu/argo-mon-status-api/pull/26) ARGO-5194 Fix the update of the tenant to not contain checks
+- [#28](https://github.com/ARGOeu/argo-mon-status-api/pull/28) ARGO-5200 FIx Upload Image
+
 
 ### Removed